项目整合mybatis及使用注解方式配置;同时使用pagehelper插件实现分页。连接数据库则使用druid

+++ cloudshopper-web/src/main/java/org/star4j/cloudshopper/controller/MainController.java
@@ -21,7 +21,7 @@ import java.util.List;
 public class MainController {
 
     @RequestMapping("/main/{id}")
-    public List<ShopAdmin> main(@PathVariable("id") long id){
+    public List<ShopAdmin> main(@PathVariable("id") Integer id){
         ShopAdmin shopAdmin = new ShopAdmin();
         shopAdmin.setAdminId(id);
         shopAdmin.setAdminName("张三");
